Experimental DSI-Studio FIB format support (https://github.com/niivue/niivue/issues/997)
